made test packages with add and subtract functions tests
