+++ .vscode/settings.json
@@ -9,7 +9,7 @@
     "**/.gitignore": false,
     "**/.vscode": true,
     "**/*.lock": true,
-    "**/node_modules": true,
+    "**/node_modules": false,
     "**/lerna-debug.log": true,
     "**/.gitmodules": true,
     "**/.cache": true
