+++ packages/pack-test-1/src/subtract/subtract.test.ts
@@ -1,0 +1,7 @@
+import { subtract } from './subtract';
+
+describe('subtract', () => {
+	it('should subtract two numbers', () => {
+		expect(subtract(5, 2)).toEqual(3);
+	});
+});
