+++ packages/pack-test-1/src/subtract/subtract.ts
@@ -1,0 +1,3 @@
+export const subtract = (a, b) => {
+	return a - b;
+};
