+++ packages/pack-test-1/src/sum/sum.test.ts
@@ -1,0 +1,7 @@
+import { sum } from './sum';
+
+describe('sum', () => {
+	it('should sum two numbers', () => {
+		expect(sum(1, 2)).toEqual(3);
+	});
+});
