+++ packages/pack-test-1/src/sum/sum.ts
@@ -1,0 +1,3 @@
+export const sum = (a, b) => {
+	return a + b;
+};
