+++ packages/pack-test-1/tsconfig.json
@@ -5,7 +5,8 @@
     "declaration": true,
     "sourceMap": true,
     "removeComments": true,
-    "allowJs": true
+    "allowJs": true,
+    "moduleResolution": "node"
   },
   "include": ["src"],
   "exclude": [
