+++ packages/pack-test-2/src/index.test.ts
@@ -1,9 +1,11 @@
-import { sum } from 'pack-test-1';
+import { ping, sum, subtract } from 'pack-test-1';
 
 test('testing works', () => {
 	expect(2 + 5).toEqual(7);
 });
 
 test('imports test correctly', () => {
+	expect(ping()).toEqual('pong');
 	expect(sum(1, 1)).toEqual(2);
+	expect(subtract(1, 1)).toEqual(2);
 });
