+++ packages/pack-test-2/src/index.ts
@@ -1,0 +1,4 @@
+import { ping, sum } from 'pack-test-1';
+
+console.log(ping());
+console.log(sum(2, 5));
